requirements.txt
gitpython==3.1.40
colorama==0.4.6 